refactor: move ms to hours fn etc to a file
