+++ front/src/features/home/Timer.js
@@ -1,20 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import useTimeDiff from './hooks/useTimeDiff'
-const second = 1000;
-const minute = 60 * second;
-const hour = minute * 60;
-
-const msToMinSec = ms => {
-  const hours = ms / hour | 0;
-  const minutes = (ms % hour) / minute | 0; // 32 bit integer shorthand
-  const seconds = (ms % minute) / second | 0;
-  return [
-    `${hours}`.padStart(2, '0'),
-    `${minutes}`.padStart(2, '0'),
-    `${seconds}`.padStart(2, '0')
-  ]
-}
+import { msToHourMinSec } from './msToHourMinSec';
 
 const renderTimer = (hours, minutes, seconds) => (
   <div className="home-timer">
@@ -25,7 +12,7 @@ const renderTimer = (hours, minutes, seconds) => (
 export default function Timer({ startDate, isActive }) {
   if (!isActive) return renderTimer('00', '00', '00');
   const diff = useTimeDiff(startDate);
-  const [hours, minutes, seconds] = msToMinSec(diff);
+  const [hours, minutes, seconds] = msToHourMinSec(diff);
   return renderTimer(hours,minutes, seconds)
 };
